decentralize jest-mock types
